refactor house
improve house component names
move button into common dir
