+++ frontend/src/Common/Buttons/index.js
@@ -1,0 +1,4 @@
+import CrossButton from './CrossButton';
+import PlusButton from './PlusButton';
+
+export {CrossButton, PlusButton};
