+++ frontend/src/House/HouseDetail.js
